ai-service/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
requests==2.31.0
langchain==0.0.308
openai>=1.0.0
anthropic==0.7.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
sentence-transformers==5.2.3
psycopg2-binary==2.9.7
pgvector==0.1.8
redis==5.0.0
pydantic==2.0.3
python-dateutil==2.8.2
gunicorn==21.2.0
